fixed some problems with hostID.

+++ framework/lib/inc/classes/host.php
@@ -41,7 +41,7 @@ class Host {
         //check, if hostID exists
         if (!file_exists(LIB_PSF_STORE . "host/hostID.php")) {
             //generate new hostID
-            $hostID = (int) uniqid(rand(), true);
+            $hostID = uniqid(rand(), true);
 
             self::$hostID = $hostID;
             $data = "<" . "?" . "php $ " . "uniqueHostID = \"" . $hostID . "\"; ?" . ">";
